[PARSER] added StubParser class
It's the old b3.fake.StubConsole but I moved it here since it's used in a production environment and if we include the b3.fake module, a FakeConsole, FakeClient and fake admin plugin will be instantiated and this can possibly use unnecessary memory
